+++ func/pvOnline.php
@@ -2,29 +2,6 @@
 
 if (!defined('pvault_panel')){ die('Not Found');}
 
-function pvOnlineStats($api, $s){
-	$jAPI = $api.'?do=getStats';
-	
-	$ch = curl_init($jAPI);
-	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-	curl_setopt($ch, CURLOPT_HTTPHEADER, array('Content-Type:application/json'));
-
-	$response = json_decode(curl_exec($ch),true);
-	curl_close($ch);
-	return $response[$s];
-	
-}
-
-function pvUploadData($pvOnlineAPI, $data){
-	$ch = curl_init($pvOnlineAPI);
-	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-	curl_setopt($ch, CURLOPT_POSTFIELDS, $data);
-	curl_setopt( $ch, CURLOPT_HTTPHEADER, array('Content-Type:application/json'));
-
-	$response = curl_exec($ch);
-	curl_close($ch);
-	return $response;
-}
 
 function pvPost($url, $data){
     $curl = curl_init($url);