+++ releasenotes.md
@@ -1,5 +1,7 @@
 Whats New in v8.2
 --------------------------
+- Remove unused pvonline functions
+- Improve pvonline stats fetch
 - Migrate pvonline to perfumersvault.com domain
 - Main website is been moved to www.perfumersvault.com
 - Show server errors if upgrade fails